requirements.txt
Flask==2.2.2
requests==2.28.1
ptvsd==4.3.2 # Required for debugging.
Pillow==9.4.0
MarkupSafe==2.1.2